show gdn price decimals

+++ resources/views/gdns/show.blade.php
@@ -132,7 +132,7 @@
                         <div class="column m-lr-20">
                             <div class="is-size- has-text-weight-bold">
                                 {{ number_format($gdn->getPaymentInCash(), 2) }}
-                                ({{ (int) $gdn->cash_received_in_percentage }}%)
+                                ({{ (float) $gdn->cash_received_in_percentage }}%)
                             </div>
                             <div class="is-uppercase is-size-7">
                                 In Cash ({{ $gdn->company->currency }})
